spawn rsync without a shell.

Make verify_all() return an "insufficient data" error if there is not at
least one supported hash type.

Make fetch() bail out when an "Insufficient data for checksum
verification" error occurs.

Do not show the spinner and the "Calculating dependencies" message for          
installation actions when the --nodeps option has been specified.
Thanks to David Watzke <david@watzke.cz> for this patch from bug #168169.

Filter paths matching '/<built-in>$' that come from debugedit.  Thanks
to Ed Catmur for this patch from bug #165342.

As requested by solar, never show a merge list unless it's explicitly
requested by --ask, --pretend, or --tree.

As requested by solar, never show a merge list unless it's explicitly
requested by --ask, --pretend, or --tree.

Allow --verbose to trigger a merge list even without --pretend, --ask,
or --tree.  Thanks to agaffney and zlin for the suggestion.

Clean up os.environ["USERLAND"] pollution from the portage.data module
so that it doesn't interfere with repoman.

For bug #167668, prune empty ${FILESDIR}s by adding  --prune-empty-dirs
to PORTAGE_RSYNC_OPTS.  Thanks to Ulrich Mueller <ulm@kph.uni-mainz.de>
for the suggestion.



svn path=/main/branches/prefix/; revision=6084

+++ pym/portage/__init__.py
@@ -912,6 +912,10 @@ class config:
 
 			# backupenv is for calculated incremental variables.
 			self.backupenv = os.environ.copy()
+			if not local_config:
+				# Clean up pollution from portage.data so that it doesn't
+				# interfere with repoman.
+				self.backupenv.pop("USERLAND", None)
 
 			def check_var_directory(varname, var):
 				if not os.path.isdir(var):
@@ -1154,7 +1158,10 @@ class config:
 
 			self.configlist.append(os.environ.copy())
 			self.configdict["env"]=self.configlist[-1]
-
+			if not local_config:
+				# Clean up pollution from portage.data so that it doesn't
+				# interfere with repoman.
+				self.configdict["env"].pop("USERLAND", None)
 
 			# make lookuplist for loading package.*
 			self.lookuplist=self.configlist[:]
@@ -2411,6 +2418,8 @@ def fetch(myuris, mysettings, listonly=0, fetchonly=0, locks_in_subdir=".locks",
 								writemsg(("!!! Got:      %s\n" + \
 									"!!! Expected: %s\n") % \
 									(reason[1], reason[2]), noiselevel=-1)
+								if reason[0] == "Insufficient data for checksum verification":
+									return 0
 								if can_fetch and not restrict_fetch:
 									writemsg("Refetching...\n\n",
 										noiselevel=-1)
@@ -2550,6 +2559,8 @@ def fetch(myuris, mysettings, listonly=0, fetchonly=0, locks_in_subdir=".locks",
 										noiselevel=-1)
 									writemsg("!!! Got:      %s\n!!! Expected: %s\n" % \
 										(reason[1], reason[2]), noiselevel=-1)
+									if reason[0] == "Insufficient data for checksum verification":
+										return 0
 									writemsg("Removing corrupt distfile...\n", noiselevel=-1)
 									os.unlink(mysettings["DISTDIR"]+"/"+myfile)
 									fetched=0
